--获取待分析数据

+++ dips-modules/gds-cleaning/gds-cleaning-api/src/main/java/com/cloud/gds/cleaning/api/vo/DataSetVo.java
@@ -25,7 +25,7 @@ public class DataSetVo {
 	/**
 	 * 权重
 	 */
-	private double weight;
+	private Float weight;
 
 	/**
 	 * 有无同义
